requirements.txt
numpy==1.14.5
flask
pygame
pyglet
opencv-python
opencv-contrib-python
pybind11
BriCA2
###oculoenv==0.1.0
#
git+https://github.com/shionhonda/oculoenv
tensorflow
chainer
chainerrl
torch
torchvision
